+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for defining Control/Acknowledge packets.
+- C++ code generation of Control/Acknowledge packets.
 ### Changed
 - Command line syntax of `vrtpktgen` changed to require generator name as a positional argument.
 - Fixed a bug in replacing existing values in packed binary fields.
+### Deprecated
+- The packet type "command" is deprecated.
 
 ## [0.1.2] - 2019-09-23
 ### Changed